backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-cloud-storage==2.19.0
google-genai==1.53.0
pydantic-settings==2.7.1
requests==2.32.3
Pillow>=12.0.0,<13.0.0
